Create Time Series Analysis of Treasury Bill Rates
The US Treasury Bill rate is a critical indicator of the macro-economic condition. The interest rates of Treasury Bills inform the fundamentals of almost all other types of investment vehicles and thus reflect the overall market condition. When the economy slows down, the Fed lowers interest rates by purchasing Treasury securities to make borrowing more accessible and therefore stimulates economic growth; on the other hand, the interest rates are raised as a measure of controlling inflation. The objective of this study is to find an appropriate time series model to understand the variation in monthly rates for one-year Treasury Bills, and then forecast future rates based on the selected model. In this analysis, the data is collected on a monthly basis from January 1975 through October 2019 for Treasury Bills with one-year maturity, processed through transformation and differencing, and fitted with various time series models to find the best forecasting model. The analysis shows that ARIMA(4,1,5) is the best model to forecast future T-Bill rates and predict market condition for the first quarter of fiscal year 2020 to inform investment decisions for both individual and private equity fund investors.